Expanded BlockID to 32b throughout H-Store. The upper 16b (at the moment) are reserved for the AntiCacheID (ACID). ACID doubles as a level count so the first AntiCacheDB is ACID = 1. The blockID change has been made across the JNI boundary back up to the Java front end. The only place a blockID is the same is in the AntiCacheDBs because these monotonically incraesing numbers are used internal to the AntiCacheDB.

+++ src/ee/anticache/EvictedTupleAccessException.h
@@ -36,7 +36,7 @@ class ReferenceSerializeOutput;
 class EvictedTupleAccessException : public SerializableEEException {
     public:
 
-        EvictedTupleAccessException(int tableId, int numBlockIds, int16_t blockIds[], int32_t tupleKeys[]);
+        EvictedTupleAccessException(int tableId, int numBlockIds, int32_t blockIds[], int32_t tupleKeys[]);
         virtual ~EvictedTupleAccessException() {}
         
         static std::string ERROR_MSG;
@@ -47,7 +47,7 @@ class EvictedTupleAccessException : public SerializableEEException {
     private:
         const int m_tableId;
         const int m_numBlockIds;
-        const int16_t *m_blockIds;
+        const int32_t *m_blockIds;
         const int32_t *m_tupleKeys;
         const int m_partitionId;
 };
