+++ src/frontend/org/voltdb/jni/ExecutionEngineIPC.java
@@ -1489,7 +1489,7 @@ public class ExecutionEngineIPC extends ExecutionEngine {
     }
 
     @Override
-    public void antiCacheReadBlocks(Table catalog_tbl, short[] block_ids, int[] tuple_offsets) {
+    public void antiCacheReadBlocks(Table catalog_tbl, int[] block_ids, int[] tuple_offsets) {
         throw new NotImplementedException("Anti-Caching is disabled for IPC ExecutionEngine");
     }
