+++ src/frontend/org/voltdb/jni/ExecutionEngineJNI.java
@@ -803,7 +803,7 @@ public class ExecutionEngineJNI extends ExecutionEngine {
 
     
     @Override
-    public void antiCacheReadBlocks(Table catalog_tbl, short[] block_ids, int[] tuple_offsets) {
+    public void antiCacheReadBlocks(Table catalog_tbl, int[] block_ids, int[] tuple_offsets) {
         if (m_anticache == false) {
             String msg = "Trying to invoke anti-caching operation but feature is not enabled";
             throw new VoltProcedure.VoltAbortException(msg);
